requirements.txt
# Agentic RAG System Requirements
# Assignment 3: Azure OpenAI & LangGraph Implementation

# Core LangGraph and ML libraries
langraph>=0.2.0
azure-ai-inference>=1.0.0
pinecone-client>=3.0.0
mlflow>=2.0.0

# AI and Embedding libraries  
langchain>=0.3.0
langchain-openai>=0.2.0
# google-generativeai>=0.8.0  # Removed - using Azure OpenAI for self-critique

# Data handling and utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0

# Development and testing
pytest>=8.0.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: For enhanced debugging
rich>=13.0.0
pretty-print>=0.1.0
